+++ scripts sans gui/liste-dc.txt
@@ -1,8 +1,2 @@
-Aquaman
 Batman
-DC Nation
-Deathstroke
-Green Arrow
-Injustice
-Super Sons
-Superman
+Naomi
